requirements.txt
torch==2.3.1
torchvision==0.18.1
ultralytics==8.2.45
opencv-python==4.10.0.84
albumentations==1.4.8
numpy==2.1.0
pandas==2.2.2
scikit-learn==1.5.1
lapx==0.5.6
filterpy==1.4.5
seaborn==0.13.2
matplotlib==3.9.1
tqdm==4.66.4
pyyaml==6.0.2
hydra-core==1.3.2
wandb==0.17.4
geopy==2.4.1
shapely==2.0.4
folium==0.17.0
fastapi==0.111.0
uvicorn==0.30.1

